Crate echo

Functions

echo
Print given strings to standard output. If scape true, it also prints the scape codes inside strings. If no_newline true, it does not print a newline after.

main

parse_code
Parse a input code from base code to a UTF-8 char. The max_digits limits how many digits the input code can have.

print_escape
Print the scape codes from string. output is where it is going to be printed.