Echo::Queue::StealingQueue

Trait Prioritized

pub trait Prioritized {
    type Kind: PartialEq + Eq + Copy;

    // Required method
    fn Rank(&self) -> Self::Kind;
}

Defines a contract for types that can be prioritized by the queue.

Required Associated Types

type Kind: PartialEq + Eq + Copy

The type of the priority value used by the implementor.

Required Methods

fn Rank(&self) -> Self::Kind

A method to retrieve the priority of the item.