Echo::Scheduler::SchedulerBuilder

Enum Concurrency

pub enum Concurrency {
    Limit(usize),
    Unlimited,
}

Defines concurrency limits for named queues (for future use).

Variants

Limit(usize)

Specifies a maximum number of concurrent tasks for a queue.

Unlimited

Allows an unlimited number of concurrent tasks for a queue.