Function echo::echo

fn echo(strings: &[String], escape: bool, no_newline: bool) -> Result<()>

Print given strings to standard output. If scape true, it also prints the scape codes inside strings. If no_newline true, it does not print a newline after.